requirements.txt
pandas
pandas-market-calendars
yfinance
pyarrow
python-dotenv
python-arango